feat: Wave 7功能冲刺—DataComparator/PacketLossDetector/ByteFrequencyAnalyzer/SerialProtocolFuzzer/FrameTemplateLibrary+25+bug修复+头文件压缩
5大新工具模块:
- DataComparator: LCS差异算法+相似度计算+Hex并排对比+CSV导出(623行)
- PacketLossDetector: 序列号追踪+间隙检测+rollover处理(465行)
- ByteFrequencyAnalyzer: Shannon熵+累积频率+模式检测(533行)
- SerialProtocolFuzzer: 6种fuzz策略+确定性PRNG+变异强度控制(678行)
- FrameTemplateLibrary: 帧模板CRUD+校验组装+JSON导入导出(633行)

25+bug修复+头文件压缩: RingBuffer线程安全+TrafficMonitor非单调时间戳+ChecksumPanel/ConverterPanel/TimestampPanel/NavIndicator/TerminalSearchRenderer/AnimatedProgressBar Doxygen+NavigationController静态变量修复

评分: 356 + 1 = 357
变更: 45 files, 3176 insertions, 57 deletions

+++ src/protocol/can/DbcParserDecode.cpp
@@ -122,10 +122,10 @@ QString DbcParser::formatSignalValue(uint32_t msgId,
     }
     if (!targetSig) { return QString(); }
 
-    /* 值表翻译(仅整数值) */
-    const int intVal = static_cast<int>(rawValue);
-    if (targetSig->valueTable.contains(intVal)) {
-        return targetSig->valueTable.value(intVal);
+    /* 值表翻译(仅整数值) — 使用qint64避免大信号值溢出 */
+    const qint64 intVal = static_cast<qint64>(rawValue);
+    if (targetSig->valueTable.contains(static_cast<int>(intVal))) {
+        return targetSig->valueTable.value(static_cast<int>(intVal));
     }
 
     /* 通用格式: "值 单位" */
